% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistDAWass-package.R
\docType{data}
\name{Agronomique}
\alias{Agronomique}
\title{Agronomique data}
\format{a \code{MatH} object, a matrix of distributions.}
\source{
Rosanna Verde, Antonio Irpino, Second University of Naples; Dominique Desbois, UMR Economie
publique, INRA-AgroParisTech, How to cope with modelling and privacy concerns? A regression model and a visualization tool for
aggregated data,  Conference of European Statistics Stakeholders,
Rome, November, 24-25,2014
}
\description{
A dataset with the distributions of marginal costs of farms in 22 France regions. It contains four 
histogram variables: "Y_TSC" (Total costs of a farm),  "X_Wheat" (Costs for Wheat), "X_Pig" (Costs for Pigs)
"X_Cmilk" (Costs for Cow Milk)
}
\author{
Antonio Irpino, 2014-10-05
}
