% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistDAWass-package.R
\docType{data}
\name{BLOOD}
\alias{BLOOD}
\title{Blood dataset for Histogram data analysis}
\format{a \code{MatH} istance, 1 row per group.}
\source{
Billard L. and Diday E. (2006). Symbolic Data Analysis: Conceptual
Statistics and Data Mining, Wiley.
}
\description{
The dataset contains a MatH (matrix of histogram-valued data) object This
data set list 14 groups of patients described by 3 variables.
}
\author{
Antonio Irpino, 2014-10-05
}
