% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistDAWass-package.R
\docType{data}
\name{BloodBRITO}
\alias{BloodBRITO}
\title{Blood dataset from Brito P. for Histogram data analysis}
\format{a \code{MatH} istance, 1 row per patient.}
\source{
Dias, S. and Brito P. Distribution and Symmetric Distribution
Regression Model for Histogram-Valued Variables, ArXiv, arXiv:1303.6199
[stat.ME]
}
\description{
The dataset contains a MatH (matrix of histogram-valued data) object This
data set list 10 patients described by 2 variables.
}
\author{
Antonio Irpino, 2014-10-05
}
