% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\name{Center.cell.MatH}
\alias{Center.cell.MatH}
\alias{Center.cell.MatH,MatH-method}
\title{Method Center.cell.MatH Centers all the cells of a matrix of distributions}
\usage{
Center.cell.MatH(object)

\S4method{Center.cell.MatH}{MatH}(object)
}
\arguments{
\item{object}{a MatH object, a matrix of distributions.}
}
\value{
A \code{MatH} object, having each distribution with a zero mean.
}
\description{
The function transform a MatH object (i.e. a matrix of distributions), 
such that each distribution is shifted and has a mean equal to zero
}
\examples{
CEN_BLOOD=Center.cell.MatH(BLOOD)
get.MatH.stats(BLOOD, stat="mean")
}
