% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_distributionH.R
\name{compP}
\alias{compP}
\alias{compP,distributionH,numeric-method}
\alias{compP,distributionH-method}
\title{Method \code{compP}}
\usage{
compP(object, q)

\S4method{compP}{distributionH,numeric}(object, q)
}
\arguments{
\item{object}{is an object of \env{distributionH} class}

\item{q}{is a numeric value}
}
\value{
Returns a value between 0 and 1.
}
\description{
Compute the cdf probability at a given value for a histogram
}
\examples{

##---- A mydist distribution ----
mydist<-distributionH(x=c(1,2,3,10), p=c(0,0.1,0.5,1))
##---- Compute the cfd value for q=5 (not observed) ----
p<-compP(mydist,5)

}
\keyword{distribution}
