% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{DouglasPeucker}
\alias{DouglasPeucker}
\title{Ramer-Douglas-Peucker algorithm for curve fitting with a PolyLine}
\usage{
DouglasPeucker(points, epsilon)
}
\arguments{
\item{points}{a 2D matrix with the coordinates of 2D points}

\item{epsilon}{an number between 0 and 1. Recomended 0.01.}
}
\value{
A matrix with the points of segments of a Poly Line.
}
\description{
Ramer-Douglas-Peucker algorithm for curve fitting with a PolyLine
}
\seealso{
\code{\link[HistDAWass]{data2hist}} function
}
