% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\name{[}
\alias{[}
\alias{[,MatH,ANY,ANY,ANY-method}
\alias{[,MatH-method}
\title{extract from a MatH Method [}
\usage{
\S4method{[}{MatH}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a \code{matH} object}

\item{i}{a set of integer values identifying the rows}

\item{j}{a set of integer values identifying the columns}

\item{...}{not useful}

\item{drop}{a logical value inherited from the basic method "[" but not used (default=TRUE)}
}
\value{
A \code{matH} object
}
\description{
This method overrides the "[" operator for a  \code{matH} object.
}
\examples{
D=BLOOD #the BLOOD dataset
SUB_D=BLOOD[c(1,2,5),c(1,2)]
}
