% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_distributionH.R
\name{get.distr}
\alias{get.distr}
\alias{get.distr,distributionH-method}
\title{Method \code{get.distr}: show the distribution}
\usage{
get.distr(object)

\S4method{get.distr}{distributionH}(object)
}
\arguments{
\item{object}{a \code{distributionH} object.}
}
\value{
A data frame: the first column contains the domain the second the CDF values.
}
\description{
This functon return the cumulative distribution function of a \code{distributionH} object.
}
\examples{
D=distributionH(x=c(1,2,3,4),p=c(0,0.2,0.6, 1))
get.distr(D) #a data.frame describing the CDF of D
}
