% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_distributionH.R
\name{get.histo}
\alias{get.histo}
\alias{get.histo,distributionH-method}
\title{Method \code{get.histo}: show the distribution with bins}
\usage{
get.histo(object)

\S4method{get.histo}{distributionH}(object)
}
\arguments{
\item{object}{a \code{distributionH} object.}
}
\value{
A matrix: the two columns contains the bounds of the histogram the third contains the probablity (or the relative frequency) of the bin.
}
\description{
This functon return a data.frame describing the histogram of a \code{distributionH} object.
}
\examples{
D=distributionH(x=c(1,2,3,4),p=c(0,0.2,0.6, 1))
get.histo(D) #returns the histogram representation of D by a data.frame
}
