% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\name{get.MatH.main.info}
\alias{get.MatH.main.info}
\alias{get.MatH.main.info,MatH-method}
\title{Method get.MatH.main.info}
\usage{
get.MatH.main.info(object)

\S4method{get.MatH.main.info}{MatH}(object)
}
\arguments{
\item{object}{a \code{MatH} object}
}
\value{
A list of char, the labels of the columns, or the names of the variables.
}
\description{
It returns the number of rows, of columns the labels of rows and columns of a \code{MatH} object.
}
\section{Slots}{

\describe{
\item{\code{nrows}}{- the number of rows}

\item{\code{ncols}}{- the number of columns}

\item{\code{rownames}}{- a vector of char, the names of rows}

\item{\code{varnames}}{- a vector of char, the names of columns}
}}

