% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\name{get.MatH.ncols}
\alias{get.MatH.ncols}
\alias{get.MatH.ncols,MatH-method}
\title{Method get.MatH.ncols}
\usage{
get.MatH.ncols(object)

\S4method{get.MatH.ncols}{MatH}(object)
}
\arguments{
\item{object}{a \code{MatH} object}
}
\value{
An integer, the number of columns.
}
\description{
It returns the number of columns of a \code{MatH} object
}
