% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\name{get.MatH.nrows}
\alias{get.MatH.nrows}
\alias{get.MatH.nrows,MatH-method}
\title{Method get.MatH.nrows}
\usage{
\S4method{get.MatH.nrows}{MatH}(object)
}
\arguments{
\item{object}{a \code{MatH} object}
}
\value{
An integer, the number of rows.
}
\description{
It returns the number of rows of a \code{MatH} object
}
