% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\name{get.MatH.rownames}
\alias{get.MatH.rownames}
\alias{get.MatH.rownames,MatH-method}
\title{Method get.MatH.rownames}
\usage{
get.MatH.rownames(object)

\S4method{get.MatH.rownames}{MatH}(object)
}
\arguments{
\item{object}{a \code{MatH} object}
}
\value{
A vector of char, the label of the rows.
}
\description{
It returns the labels of the rows of a \code{MatH} object
}
