% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\name{get.MatH.varnames}
\alias{get.MatH.varnames}
\alias{get.MatH.varnames,MatH-method}
\title{Method get.MatH.varnames}
\usage{
get.MatH.varnames(object)

\S4method{get.MatH.varnames}{MatH}(object)
}
\arguments{
\item{object}{a \code{MatH} object}
}
\value{
A vector of char, the labels of the columns, or the names of the variables.
}
\description{
It returns the labels of the columns, or the names of the variables, of a \code{MatH} object
}
