% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_classes.R
\docType{class}
\name{HTS-class}
\alias{HTS-class}
\alias{HTS}
\alias{initialize,HTS-method}
\title{Class HTS}
\usage{
\S4method{initialize}{HTS}(.Object, epocs = 1, ListOfTimedElements = c(new("TdistributionH")))
}
\arguments{
\item{.Object}{the object type ("HTS") a histogram time series}

\item{epocs}{the number of histograms (one for each timepoint or period)}

\item{ListOfTimedElements}{a vector of \code{TdistributionH} objects}
}
\description{
Class \code{HTS} defines a histogram time series, i.e. a set of histograms observed along time
}
