% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_time_series.R
\name{HTS.moving.averages}
\alias{HTS.moving.averages}
\title{Smoothing with moving averages of a histogram time series}
\usage{
HTS.moving.averages(HTS, k = 3, weights = rep(1, k))
}
\arguments{
\item{HTS}{A \code{HTS} object (a histogram time series).}

\item{k}{an integer value, the number of elements for moving averages}

\item{weights}{a vector of positive weights for a weighted moving average}
}
\value{
a list with the results of the smoothing procedure.
}
\description{
(Beta verson of) Extends the moving average smoothing of a time 
series to a histogram time series, using L2 Wasserstein distance.
}
\section{Slots}{

\describe{
\item{\code{k}}{the number of elements for the average}

\item{\code{weights}}{the vector of weights for smoothing}

\item{\code{AveragedHTS}}{The smoothed HTS}
}}

\examples{
mov.av.smoothed=HTS.moving.averages(HTS=RetHTS, k=5)
# a show method for HTS must be implemented you can see it using
# str(mov.av.smoothed$AveragedHTS)
}
