% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_distributionH.R
\name{meanH}
\alias{meanH}
\alias{meanH,distributionH-method}
\title{Method \code{meanH}: computes the mean of a distribution}
\usage{
meanH(object)

\S4method{meanH}{distributionH}(object)
}
\arguments{
\item{object}{a \code{distributionH} object}
}
\value{
the mean of the distribution
}
\description{
Mean of a histogram (First moment of the distribution)
}
\examples{

##---- A mydist distribution ----
mydist<-distributionH(x=c(1,2,3,10), p=c(0,0.1,0.5,1))
##---- Compute the mean of mydist ----
meanH(mydist) #---> 4.4

}
\author{
Antonio Irpino
}
\keyword{distribution}
