% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistDAWass-package.R
\docType{data}
\name{OzoneH}
\alias{OzoneH}
\title{Complete Ozone dataset for Histogram data analysis}
\format{a \code{MatH} istance, 1 row per station.}
\source{
http://java.epa.gov/castnet/epa_jsp/prepackageddata.jsp
ftp://ftp.epa.gov/castnet/data/metdata.zip
}
\description{
The dataset contains MatH (matrix of histogram-valued data) object This data
set list 84 stations located in the USA recording four variables. Some
stations contains missing data.
}
\author{
Antonio Irpino, 2014-10-05
}
