% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_distributionH.R
\docType{methods}
\name{+}
\alias{+}
\alias{+,distributionH,distributionH-method}
\alias{+,numeric,distributionH-method}
\alias{+,distributionH,numeric-method}
\title{Method +}
\usage{
\S4method{+}{distributionH,distributionH}(e1, e2)

\S4method{+}{numeric,distributionH}(e1, e2)

\S4method{+}{distributionH,numeric}(e1, e2)
}
\arguments{
\item{e1}{a \code{distributionH} object or a number}

\item{e2}{a \code{distributionH} object or a number}
}
\value{
a \code{distributionH} object
}
\description{
the sum of two distribution according to the L2 Wasssertein

the sum of a number and a distribution according to the L2 Wasssertein

the sum of adistribution and a number according to the L2 Wasssertein
}
