% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistDAWass-package.R
\docType{data}
\name{RetHTS}
\alias{RetHTS}
\title{A histogram-valued dataset of returns}
\format{a \code{MatH} object, a matrix of distributions.}
\description{
A histogram-valued dataset of returns of dollar vs yen change rates
}
\author{
Antonio Irpino, 2014-10-05
}
