% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\name{set.cell.MatH}
\alias{set.cell.MatH}
\alias{set.cell.MatH,distributionH,MatH,numeric,numeric-method}
\alias{set.cell.MatH,MatH-method}
\title{Method set.cell.MatH assign a histogram to a cell of a matrix of histograms}
\usage{
set.cell.MatH(object, mat, r, c)

\S4method{set.cell.MatH}{distributionH,MatH,numeric,numeric}(object, mat, r, c)
}
\arguments{
\item{object}{a distributionH object, a matrix of distributions.}

\item{mat}{a MatH object, a matrix of distributions.}

\item{r}{an integer, the row index.}

\item{c}{an integer, the column index}
}
\value{
A \code{MatH} object.
}
\description{
Assign a histogram data to the r-th row and the c-th column of a matrix of histograms.
}
\examples{
mydist=distributionH(x=c(0,1,2,3,4), p=c(0,0.1,0.6,0.9,1))
MAT=set.cell.MatH(mydist,BLOOD, r=1,c=1)
}
