% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{ShortestDistance}
\alias{ShortestDistance}
\title{Shortes distance from a point o a 2d segment}
\usage{
ShortestDistance(p, line)
}
\arguments{
\item{p}{coordinates of a point}

\item{line}{a 2x2 matrix with the coordinates of two points defining a line}
}
\value{
A numeric value, the Euclidean distance of point \code{p} to the \code{line}.
}
\description{
Shortes distance from a point o a 2d segment
}
\seealso{
\code{\link[HistDAWass]{data2hist}} function and \code{\link[HistDAWass]{DouglasPeucker}} function
}
