% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_distributionH.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,distributionH-method}
\title{Method show for distributionH}
\usage{
\S4method{show}{distributionH}(object)
}
\arguments{
\item{object}{a \code{distributionH} object}
}
\description{
An overriding show function for a \code{distributionH} object. The function returns a representation 
of the histogram, if the number of bins is high the central part of the histogram is truncated.
}
\examples{
##---- initialize a distributionH
 mydist<-distributionH(x=c(7,8,10,15),p=c(0, 0.2, 0.7, 1))
 # show the histogram
 mydist
}
