% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\docType{methods}
\name{show-MatH}
\alias{show-MatH}
\alias{show,MatH-method}
\title{Method show for MatH}
\usage{
\S4method{show}{MatH}(object)
}
\arguments{
\item{object}{a \code{MatH} object}
}
\description{
An overriding show method for a \code{MatH} object. The method returns a representation 
of the matrix using the mean and the standard deviation for each histogram.
}
\examples{
show(BLOOD)
print(BLOOD)
BLOOD
}
