% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistDAWass-package.R
\docType{data}
\name{stations_coordinates}
\alias{stations_coordinates}
\title{Stations coordinates of China_Month and China_Seas datasets}
\format{a data.frame}
\source{
raw data are available here: \url{http://cdiac.ess-dive.lbl.gov/ftp/tr055/}. 
Climate Data Bases of the People's Republic of China 1841-1988 (TR055)
DOI: 10.3334/CDIAC/cli.tr055
}
\description{
A dataset containing the geographical coordinates of stations described in China_Month and China_Seas datasets
}
\author{
Antonio Irpino, 2014-10-05
}
