% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_distributionH.R
\name{stdH}
\alias{stdH}
\alias{stdH,distributionH-method}
\title{Method \code{stdH}: computes the standard deviation of a distribution}
\usage{
stdH(object)

\S4method{stdH}{distributionH}(object)
}
\arguments{
\item{object}{a \code{distributionH} object}
}
\value{
A value for the standard deviation
}
\description{
Standard deviation of a histogram (i.e., the square root of the centered
second moment)
}
\examples{

##---- A mydist distribution ----
mydist<-distributionH(x=c(1,2,3,10), p=c(0,0.1,0.5,1))
##---- Compute the standard deviation of mydist ----
stdH(mydist) #---> 2.563851

}
\author{
Antonio Irpino
}
\keyword{distribution}
