% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_HTS.R
\name{subsetHTS}
\alias{subsetHTS}
\alias{subsetHTS,HTS,numeric,numeric-method}
\title{Method \code{subsetHTS}: extract a subset of a histogram time series}
\usage{
subsetHTS(object, from, to)

\S4method{subsetHTS}{HTS,numeric,numeric}(object, from, to)
}
\arguments{
\item{object}{a \code{HTS} object. A histogram 1d time series}

\item{from}{an integer, the initioal timepont}

\item{to}{an integer, a final timepoint}
}
\value{
a \code{HTS} object. A histogram 1d time series
}
\description{
This functon return the mean of a \code{distributionH} object.
}
\examples{
SUB_RetHTS=subsetHTS(RetHTS,from=1, to=20)# the first 20 elements
}
