% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_with_ggplot.R
\name{summaryHTS}
\alias{summaryHTS}
\title{A function for summarize HTS}
\usage{
summaryHTS(x)
}
\arguments{
\item{x}{a HTS}
}
\value{
A matrix with basic statistics.
}
\description{
A summarizer for HTS
}
\examples{
summaryHTS(subsetHTS(RetHTS, from = 1, to = 10))
}
