% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_classes.R
\docType{class}
\name{TdistributionH-class}
\alias{TdistributionH-class}
\alias{TdistributionH}
\alias{initialize,TdistributionH-method}
\title{Class TdistributionH}
\usage{
\S4method{initialize}{TdistributionH}(
  .Object,
  tstamp = numeric(0),
  period = list(start = -Inf, end = -Inf),
  x = numeric(0),
  p = numeric(0),
  m = numeric(0),
  s = numeric(0)
)
}
\arguments{
\item{.Object}{the type of object ("TdistributionH") a \code{"distributionH"} object with a time reference}

\item{tstamp}{a numeric value related to  a timestamp}

\item{period}{a list of two values, the starting time and the ending time (alternative to tstamp if the
distribution is observed along a period and not on a timestamp)}

\item{x}{a vector of increasing values, the domain of the distribution (the same of \code{distributionH} object)}

\item{p}{a vector of increasing values from 0 to 1, 
the CDF of the distribution (the same of \code{distributionH} object)}

\item{m}{a number, the mean of the distribution (the same of \code{distributionH} object)}

\item{s}{a positive number, the standard deviation of the distribution (the same of \code{distributionH} object)}
}
\description{
Class \code{TdistributionH} defines a histogram with a time (point or period)
}
