% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_classes.R
\docType{class}
\name{TMatH-class}
\alias{TMatH-class}
\alias{TMatH}
\alias{initialize,TMatH-method}
\title{Class TMatH}
\usage{
\S4method{initialize}{TMatH}(
  .Object,
  tstamp = numeric(0),
  period = list(start = -Inf, end = -Inf),
  mat = new("MatH")
)
}
\arguments{
\item{.Object}{the type of object ("TMatH")}

\item{tstamp}{a vector of time stamps, numeric.}

\item{period}{a list of pairs with a vectorof starting time and a vector ofending time.
This parameter is used alternatively to \code{tstamp} if the distributions are related to time periods
instead of timestamps}

\item{mat}{a \code{MatH} object}
}
\description{
Class \code{TMatH} defines a matrix of histograms, a \code{TMatH} object, with a time (a timepoint or a time window).
}
