% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\name{WH.bind.row}
\alias{WH.bind.row}
\alias{WH.bind.row,MatH,MatH-method}
\alias{WH.bind.row,MatH-method}
\title{Method WH.bind.row}
\usage{
WH.bind.row(object1, object2)

\S4method{WH.bind.row}{MatH,MatH}(object1, object2)
}
\arguments{
\item{object1}{a \code{MatH} object}

\item{object2}{a \code{MatH} object}
}
\value{
a \code{MatH} object,
}
\description{
It attaches two \code{MatH} objects with the same columns by row.
}
\examples{
M1<-BLOOD[1:3,]
M2<-BLOOD[5:8,]
MAT<-WH.bind.row(M1,M2)
}
