% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met_MatH.R
\name{WH.mat.sum}
\alias{WH.mat.sum}
\alias{WH.mat.sum,MatH,MatH-method}
\alias{WH.mat.sum,MatH-method}
\title{Method WH.mat.sum}
\usage{
WH.mat.sum(object1, object2)

\S4method{WH.mat.sum}{MatH,MatH}(object1, object2)
}
\arguments{
\item{object1}{a \code{MatH} object}

\item{object2}{a \code{MatH} object}
}
\value{
a \code{MatH} object,
}
\description{
It sums two \code{MatH} objects, i.e. two matrices of distributions, 
by summing the quantile functions of histograms. This sum is consistent with 
a set of distributions equipped with a L2 wasserstein metric.
}
\examples{
# binding by row 
M1<-BLOOD[1:5,]
M2<-BLOOD[6:10,]
MAT<-WH.mat.sum(M1,M2)
}
