% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Met_MatH.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,MatH,ANY,ANY,ANY-method}
\title{Method [}
\usage{
\S4method{[}{MatH,ANY,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{a \code{matH} object}

\item{i}{a set of integer values identifying the rows}

\item{j}{a set of integer values identifying the columns}

\item{drop}{a logical value iherited from the basic method "[" but not used (default=TRUE)}

\item{...}{not useful}
}
\value{
A \code{matH} object
}
\description{
This method overrides the "[" operator for a  \code{matH} object.
}
\examples{
D=BLOOD #the BLOOD dataset
SUB_D=BLOOD[c(1,2,5),c(1,2)]
}

