% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Met_distributionH.R
\docType{methods}
\name{get.m}
\alias{get.m}
\alias{get.m,distributionH-method}
\title{Method \code{get.m}: the mean of a distribution}
\usage{
get.m(object)

\S4method{get.m}{distributionH}(object)
}
\arguments{
\item{object}{a \code{distributionH} object}
}
\value{
A numeric value
}
\description{
Method \code{get.m}: the mean of a distribution

This functon return the mean of a \code{distributionH} object.
}
\examples{
D=distributionH(x=c(1,2,3,4),p=c(0,0.2,0.6, 1))
get.m(D) #returns the mean of D
}

