% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Met_MatH.R
\docType{methods}
\name{plot-MatH}
\alias{plot,MatH-method}
\alias{plot-MatH}
\title{Method plot for a matrix  of histograms}
\usage{
\S4method{plot}{MatH}(x, y = "missing", type = "HISTO", col = "green",
  border = "black")
}
\arguments{
\item{x}{a \code{distributionH} object}

\item{y}{not used in this implementation}

\item{type}{(optional) a string describing the type of plot, default="HISTO".\cr
Other allowed types are \cr
"DENS"=a density approximation, \cr
"BOXPLOT"=l boxplot}

\item{col}{(optional) a string the color of the plot, default="green".}

\item{border}{(optional) a string the color of the border of the plot, default="black".}
}
\description{
An overloading plot function for a \code{MatH} object. The method returns a graphical representation
of the matrix of histograms.
}
\examples{
plot(BLOOD) #plots BLOOD dataset
 plot(BLOOD, type="HISTO", col="red", border="blue") #plots a matrix of histograms
 plot(BLOOD, type="DENS", col="red", border="blue") #plots a matrix of densities
 plot(BLOOD, type="BOXPLOT") #plots a  boxplots
}

