% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Met_distributionH.R
\docType{methods}
\name{skewH}
\alias{skewH}
\alias{skewH,distributionH-method}
\title{Method \code{skewH}: computes the skewness of a distribution}
\usage{
skewH(object)

\S4method{skewH}{distributionH}(object)
}
\arguments{
\item{object}{a \code{distributionH} object}
}
\value{
A value for the skewness index
}
\description{
Method \code{skewH}: computes the skewness of a distribution

Skewness of a histogram (using the third standardized moment)
}
\examples{
##---- A mydist distribution ----
mydist<-distributionH(x=c(1,2,3,10), p=c(0,0.1,0.5,1))
##---- Compute the skewness of mydist ----
skewH(mydist) #---> -1.186017
}
\author{
Antonio Irpino
}
\keyword{distribution}

