% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Met_MatH.R
\docType{methods}
\name{WH.bind.col}
\alias{WH.bind.col}
\alias{WH.bind.col,MatH,MatH-method}
\alias{WH.bind.col,MatH-method}
\title{Method WH.bind.col}
\usage{
WH.bind.col(object1, object2)

\S4method{WH.bind.col}{MatH,MatH}(object1, object2)
}
\arguments{
\item{object1}{a \code{MatH} object}

\item{object2}{a \code{MatH} object}
}
\value{
a \code{MatH} object,
}
\description{
Method WH.bind.col

It attaches two \code{MatH} objects with the same rows by colums.
}
\examples{
M1<-BLOOD[1:10,1]
M2<-BLOOD[1:10,3]
MAT<-WH.bind.col(M1,M2)
}

