% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_components.R
\name{WH.MultiplePCA}
\alias{WH.MultiplePCA}
\title{(Preliminary version) Principal components analysis  of a set of histogram variable based on Wasserstein distance}
\usage{
WH.MultiplePCA(data, list.of.vars, quantiles = 10)
}
\arguments{
\item{data}{A MatH object (a matrix of distributionH).}

\item{list.of.vars}{A list of  integers, the active variables.}

\item{quantiles}{An integer, it is the number of quantiles used in the analysis.
Default=10, all the densities are considered.}
}
\value{
a list with the results of the PCA in the MFA format of package \pkg{FactoMineR} for function MFA
}
\description{
(Beta version) The function implements a Principal components analysis of a set of histogram variables 
based on Wasserstein distance. It performs a centered (not standardized) PCA on a set of quantiles of a variable.
Being a distribution a multivalued description, the analysis performs a dimensional reduction and a visualization of distributions. 
It is a 1d (one dimension) becuse it is considered just one histogram variable.
}
\details{
It is a first tentative of extending WH.1d.PCA to the multiple case.All plotting functions must be defined
in order to have interpretable results. This is an ongoing function, even if the numerical results are methodologically verified.
}
